backend/requirements.txt
# ===========================
# SOR-HD Backend Requirements
# ===========================

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Background Tasks (Optional - for async optimization)
celery==5.3.6
redis==5.0.1

# Route Optimization
ortools==9.14.6206

# Notifications
# Firebase Cloud Messaging (FCM) for Android push notifications
firebase-admin==6.4.0
# Apple Push Notification Service (APNS) for iOS push notifications
aioapns==3.2
# Twilio for SMS notifications
twilio==8.11.1

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
shapely==2.0.6  # Updated for numpy 2.x compatibility
numpy<2.0.0  # Ensure compatibility with shapely
